refactor: consolidate utility functions into dedicated files
