+++ src/actions/saveContact.ts
@@ -15,9 +15,7 @@ import {
   FlattenContact,
   PlainFields,
 } from '@/lib/definitions'
-import { dateString } from '@/lib/utils'
 import { Address, Email, Occupation, Organization } from '@prisma/client'
-import { useSession } from 'next-auth/react'
 import { Session } from 'next-auth/types'
 
 export type State = {
